% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{arrayLift}
\alias{arrayLift}
\title{arrayLift}
\usage{
arrayLift(probes = probes, array = "EPIC")
}
\arguments{
\item{probes}{A dataframe or vector of EPIC, 450K, or 27K CpG IDs}

\item{array}{A character with array platform ("EPIC", "450K" or "27K")}
}
\value{
A \code{GRanges} object of hg38 coordinates
}
\description{
LiftOver EPIC, 450K, or 27K infinium array CpG IDs to hg38 coordinates
}
\examples{
\dontrun{ 
readxl::read_excel("file.xlsx") \%>\%
  dplyr::select(CpGID) \%>\% 
  arrayLift("EPIC") \%>\%
  dplyr::as_tibble() \%>\%
  dplyr::select(seqnames, start, end) \%>\% 
  DMRichR::df2bed("file.bed")
}
}
\references{
\url{https://support.bioconductor.org/p/78652/}
}
