% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{CCplot}
\alias{CCplot}
\title{CCplot}
\usage{
CCplot(
  tidyCC = tidyCC,
  testCovariate = testCovariate,
  adjustCovariate = NULL,
  matchCovariate = NULL
)
}
\arguments{
\item{tidyCC}{The list of tibbles returned by \code{DMRichR::CCstats()}}

\item{testCovariate}{The factor to test for differences between groups}

\item{adjustCovariate}{The covariate(s) to adjust for between groups}

\item{matchCovariate}{Another covariate to adjust for between groups (for dmrseq compatibility)}

\item{cellComposition}{Character vector of the cell composition dataset}
}
\value{
A \code{ggplot} object that can be viewed by calling it, saved with \code{ggplot2::ggsave()},
 or further modified by adding \code{ggplot2} syntax.
}
\description{
Plots the cell composition differences between groups.
}
