% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{CCstats}
\alias{CCstats}
\title{CCstats}
\usage{
CCstats(
  CC = NULL,
  bs.filtered.bsseq.cc = bs.filtered.bsseq.cc,
  testCovariate = testCovariate,
  adjustCovariate = NULL,
  matchCovariate = NULL
)
}
\arguments{
\item{CC}{A \code{matrix} from \code{DMRichR::Houseman()}}

\item{bs.filtered.bsseq.cc}{Smoothed \code{bsseq} object with design matrix in pData}

\item{testCovariate}{The factor to test for differences between groups}

\item{adjustCovariate}{The covariate(s) to adjust for between groups}

\item{matchCovariate}{Another covariate to adjust for between groups (for dmrseq compatibility)}
}
\value{
A list of tibbles with the statsitics and the values used for the tests
}
\description{
Computes the cell composition differences between groups while adjusting for the provided covariates. 
 The differences are tested for using an ANOVA through the \code{\link[stats]{aov}} function.
}
\references{
\url{https://cran.r-project.org/web/packages/broom/vignettes/broom_and_dplyr.html}
}
