% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalPlots.R
\name{CGi}
\alias{CGi}
\title{Extract methylation values from CpG islands}
\usage{
CGi(bs.filtered.bsseq = bs.filtered.bsseq, genome = genome)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}

\item{genome}{A character vector of the genome of interest (i.e. "hg38")}
}
\value{
A matrix of smoothed individual methylation values
}
\description{
Obtain individual smoothed methylation values for CpG islands
}
\details{
CGi
}
