% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{CGiPCA}
\alias{CGiPCA}
\title{CGiPCA}
\usage{
CGiPCA(bsseq = bs.filtered.bsseq, genome = genome, group = NA)
}
\arguments{
\item{bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}

\item{genome}{A character vector of the genome of interest (i.e. "hg38")}
}
\value{
A \code{ggplot} object that can be viewed by calling it,
 saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Performs and plots a PCA of CpG island from individual smoothed methylation values
}
