% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOLA.R
\name{chromHMM}
\alias{chromHMM}
\title{Chromatin state enrichments}
\usage{
chromHMM(sigRegions = sigRegions, regions = regions, cores = cores)
}
\arguments{
\item{sigRegions}{A \code{GRanges} object of significant regions}

\item{regions}{A \code{GRanges} object of background regions}

\item{cores}{An integer of how many cores to use}
}
\value{
A \code{tibble} of enrichment results
}
\description{
Perfom enrichment testing against the ChromHMM 15-state model for hg38
 using \code{LOLA}
}
\details{
chromHMM
}
