% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOLA.R
\name{chromHMM_heatmap}
\alias{chromHMM_heatmap}
\title{Chromatin state heatmap}
\usage{
chromHMM_heatmap(chromHMM = chromHMM)
}
\arguments{
\item{chromHMM}{A \code{tibble} of enrichment results}
}
\value{
Saves a heatmap
}
\description{
Plot a heatmap of \code{LOLA} enrichment testing results of the ChromHMM
 15-state model for hg38
}
\details{
chromHMM_heatmap
}
