% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanRanges}
\alias{cleanRanges}
\title{cleanRanges}
\usage{
cleanRanges(gr = gr)
}
\arguments{
\item{gr}{Genomic ranges or bsseq object}
}
\value{
Genomic ranges or bsseq object with junk contigs and the mitochondrial chromosome removed
}
\description{
Remove junk contigs and the mitochondrial chromosome from a genomic ranges or bsseq object.
}
