% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{CpGs}
\alias{CpGs}
\title{CpGs}
\usage{
CpGs(bs.filtered.bsseq = bs.filtered.bsseq)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}
}
\value{
A matrix of smoothed individual methylation values
}
\description{
Performs and plots a PCA of single CpG individual smoothed methylation values
}
