% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{df2bed}
\alias{df2bed}
\title{df2bed}
\usage{
df2bed(df = df, bed = bed)
}
\arguments{
\item{df}{Data frame}

\item{bed}{Name of the bed file in quotations}
}
\value{
Bed file
}
\description{
Save a dataframe as a basic bed file
}
