% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{DM.R}
\alias{DM.R}
\title{Run the pipeline}
\usage{
DM.R(
  genome = c("hg38", "hg19", "mm10", "mm9", "rheMac10", "rheMac8", "rn6", "danRer11",
    "galGal6", "bosTau9", "panTro6", "dm6", "susScr11", "canFam3", "TAIR10", "TAIR9"),
  coverage = 1,
  perGroup = 1,
  minCpGs = 5,
  maxPerms = 10,
  maxBlockPerms = 10,
  cutoff = 0.05,
  testCovariate = testCovariate,
  adjustCovariate = NULL,
  matchCovariate = NULL,
  cores = 20,
  GOfuncR = TRUE,
  sexCheck = FALSE,
  EnsDb = FALSE,
  cellComposition = FALSE
)
}
\arguments{
\item{genome}{Character specifying the genome.}

\item{coverage}{CpG coverage cutoff (1x recommended)}

\item{perGroup}{Percent of samples per a group to apply the CpG coverage cutoff to (from 0 to 1)}

\item{testCovariate}{Factor of interest. Only case vs control is supported.}

\item{adjustCovariate}{Variables to adjust for. Multiple variables can be provided.}

\item{matchCovariate}{Variable to block for when constructing permutations. 
Only a single variable can be provided and it cannot also be an adjustCovariate.}

\item{cores}{Integer specifying the number of cores to use. 20 is recommended.}

\item{GOfuncR}{Logical indicating whether to run a GOfuncR GO analysis.}

\item{sexCheck}{Logical indicating whether to confirm sex of each sample. 
This is highly recommended if your analysis has males and females and will also drop the sex chromosomes.}

\item{EnsDb}{Logical indicating whether to to select Ensembl transcript annotation database.
This is reccomended for non-model organisms.}

\item{cellComposition}{Logical indicating whether to estimate whole-blood cell composition.
This is an experimental argument.}

\item{meta}{Design matrix data frame with sample name in the Name column}
}
\description{
Performs the entire DMRichR analysis pipeline.
}
\details{
DM.R
}
