% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_cgmaptools.R
\name{DM_cgmaptools.R}
\alias{DM_cgmaptools.R}
\title{Run the pipeline for dmr derived from CGmaptools with single replicate}
\usage{
DM_cgmaptools.R(
  genome = c("hg38", "hg19", "mm10", "mm9", "rheMac10", "rheMac8", "rn6", "danRer11",
    "galGal6", "bosTau9", "panTro6", "dm6", "susScr11", "canFam3", "TAIR10", "TAIR9"),
  minSites = 5,
  cutoff = 0.05,
  cores = 20,
  GOfuncR = TRUE,
  fileName = "*_dmr.CG.txt.gz",
  EnsDb = FALSE,
  internet = TRUE
)
}
\arguments{
\item{genome}{Character specifying the genome.}

\item{minSites}{Numeric for minimum number of cytosine sites for a DMR.}

\item{cutoff}{Numeric cutoff for qvalues}

\item{cores}{Numeric specifying the number of cores to use. 20 is recommended.}

\item{GOfuncR}{Logical indicating whether to run a GOfuncR GO analysis.}

\item{fileName}{Character indicating dmr file name pattern}

\item{EnsDb}{Whether to use Ensemble database}

\item{internet}{Logical indicating if internet connection is available}
}
\description{
Performs the entire DMRichR analysis pipeline, 
using a limited number of tools.
}
\details{
DM_cgmaptools.R
}
