% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRich.R
\name{DMRichGenic}
\alias{DMRichGenic}
\title{Gene region enrichment testing}
\usage{
DMRichGenic(
  sigRegions = sigRegions,
  regions = regions,
  TxDb = TxDb,
  annoDb = annoDb
)
}
\arguments{
\item{sigRegions}{\code{GRanges} object of DMRs}

\item{regions}{\code{GRanges} object of background regions}

\item{TxDb}{TxDb annotation package for genome of interest}

\item{annoDb}{Character specifying OrgDb annotation package for species of interest}
}
\value{
A tibble with the enrichment results
}
\description{
Test DMRs for overlaps with gene region annotations for all genomes
}
\details{
DMRichGenic
}
