% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRich.R
\name{DMRichGenicPlot}
\alias{DMRichGenicPlot}
\title{DMRichGenicPlot}
\usage{
DMRichGenicPlot(genicOverlaps = genicOverlaps)
}
\arguments{
\item{genicOverlaps}{\code{tibble} from \code{DMRichR::DMRichGenic}.}
}
\value{
A \code{ggplot} object of enrichment results that can be viewed by calling it, 
saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Plot DMR gene region enrichment testing results from \code{DMRichR::DMRichGenic}.
}
