% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOLA.R
\name{dmrList}
\alias{dmrList}
\title{dmrList}
\usage{
dmrList(sigRegions = sigRegions)
}
\arguments{
\item{sigRegions}{A \code{GRanges} object of DMRs from \code{dmrseq::dmrseq()}}
}
\value{
A \code{GRangesList} of DMRs
}
\description{
Create \code{GRangesList} object of all, hypermethylated, and hypomethylated DMRs from \code{dmrseq::dmrseq()}
}
