% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRichR-package.R
\name{.onAttach}
\alias{.onAttach}
\title{onAttach}
\usage{
.onAttach(libname = find.package("DMRichR"), pkgname = "DMRichR")
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
onAttach
}
