% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS.R
\name{DSS.R}
\alias{DSS.R}
\title{Run the pipeline with DSS as the DMR calling method}
\usage{
DSS.R(
  genome = c("hg38", "hg19", "mm10", "mm9", "rheMac10", "rheMac8", "rn6", "danRer11",
    "galGal6", "bosTau9", "panTro6", "dm6", "susScr11", "canFam3", "TAIR10", "TAIR9"),
  analysisType = "twoGroup",
  condition1 = "",
  condition2 = "",
  minDiff = 0.1,
  pval_cutoff = 0.05,
  minSites = 3,
  ratio_cutoff = 2,
  factor1 = "",
  factor2 = "",
  ref1 = "",
  ref2 = "",
  cores = 10,
  wd = ".",
  resPath = NULL,
  override = FALSE
)
}
\arguments{
\item{genome}{Character specifying the genome.}

\item{analysisType}{Character indicating type of analysis design, either 'general' or 'twoGroup'.}

\item{condition1}{Character indicating the group1.}

\item{condition2}{Character indicating the group2.}

\item{pval_cutoff}{Numeric cutoff value [from 0 to 1] for the pval of DML used for DMR detection.}

\item{minSites}{Numeric for the minimum number of Cytosines for a DMR.}

\item{ratio_cutoff}{cutoff value [from 0 to inf] for the ratio areaStat/nSites used for DMR detection.}

\item{factor1}{Character indicating factor of interest from the design matrix.}

\item{factor2}{Character indicating co-factor of interest from the design matrix.}

\item{ref1}{Character indicating reference condition for the factor of interest from the design matrix.}

\item{ref2}{Character indicating reference condition for the co-factor of interest from the design matrix.}

\item{cores}{Numeric specifying the number of cores to use. 10 is recommended.}

\item{wd}{Character indicating the location where the analysis results to be stored.}

\item{resPath}{character specifying path to local resources if internet is not available.}

\item{override}{logical indicating whether to redefine DMRs, default FALSE.}

\item{minDifff}{cutoff value [from 0 to inf] for the minimum difference between mean methylation 
levels between group1 and group2 during DMR detection for twoGroup analysis.}
}
\description{
Performs the entire DMRichR analysis pipeline, 
which runs most functions in the package.
}
\details{
DSS.R
}
