% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS_specific.R
\name{DSS_multi_factor}
\alias{DSS_multi_factor}
\title{DML and DMR calling with DSS method}
\usage{
DSS_multi_factor(
  bss,
  desgin,
  factor1,
  factor2,
  pval_cutoff,
  ratio_cutoff,
  minSites = 3
)
}
\arguments{
\item{bss}{a smoothed bsseq object.}

\item{factor1}{Character indicating factor of interest from the design matrix.}

\item{factor2}{Character indicating co-factor of interest from the design matrix.}

\item{pval_cutoff}{Numeric cutoff value [from 0 to 1] for the pval of DML used for DMR detection.}

\item{ratio_cutoff}{cutoff value [from 0 to inf] for the ratio areaStat/nSites used for DMR detection.}

\item{minSites}{Numeric for the minimum number of Cytosines for a DMR.}

\item{design}{sample information dataframe.}
}
\value{
a list of DMRs for all factors and interaction. See \code{findDMR} for details
}
\description{
Call DML using multifactor design, then get DMRS for each factor 
and interaction effect if two factors are provided in the design
}
\details{
DSS_multi_factor
}
