% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS_specific.R
\name{DSS_pairwise}
\alias{DSS_pairwise}
\title{DML and DMR calling with DSS method for two groups}
\usage{
DSS_pairwise(bss, condition1, condition2, pval_cutoff, minDiff, minSites = 3)
}
\arguments{
\item{bss}{a bsseq object.}

\item{condition1}{Character indicating the group1.}

\item{condition2}{Character indicating the group2.}

\item{pval_cutoff}{Numeric cutoff value [from 0 to 1] for the pval of DML used for DMR detection.}

\item{minSites}{Numeric for the minimum number of Cytosines for a DMR.}

\item{minDifff}{cutoff value [from 0 to inf] for the minimum difference between mean methylation 
levels between group1 and group2 during DMR detection.}
}
\value{
a list of DMRs and background regions
}
\description{
Call DML using two group design, then get DMRS and background regions.
}
\details{
DSS_pairwise
}
