% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{extend}
\alias{extend}
\title{extend}
\usage{
extend(x, upstream = 0, downstream = 0)
}
\arguments{
\item{x}{A \code{GRanges} object to extend}

\item{upstream}{Numeric of basepairs to extend the 5' end by}

\item{downstream}{Numeric of basepairs to extend the 3' end by}
}
\value{
A \code{GRanges} object with extended ranges
}
\description{
Extend the 5' and 3' ends of a \code{GRanges}. See references for source.
}
\references{
\url{https://support.bioconductor.org/p/78652/}
}
