% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS_specific.R
\name{findDMR}
\alias{findDMR}
\title{DMR calling from DML}
\usage{
findDMR(DML, pval_cutoff = 0.05, ratio_cutoff = 2, minSites = 3)
}
\arguments{
\item{DML}{a smoothed bsseq object.}

\item{pval_cutoff}{Numeric cutoff value [from 0 to 1] for the pval of DML used for DMR detection.}

\item{ratio_cutoff}{cutoff value [from 0 to inf] for the ratio areaStat/nSites used for DMR detection.}

\item{minSites}{Numeric for the minimum number of Cytosines for a DMR.}
}
\value{
a list of two dataframes, the first is for signigicant DMRs and the second is for background DMRs
}
\description{
Call DMR using multifactor design
}
\details{
findDMR
}
