% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getBackground}
\alias{getBackground}
\title{getBackground}
\usage{
getBackground(bs = bs.filtered, minNumRegion = 5, maxGap = 1000)
}
\arguments{
\item{bs}{bsseq object that has been filtered for coverage and sorted}

\item{minNumRegion}{Minimum CpGs required for a region. Must be at least 3 CpGs, default is 5 CpGs}

\item{maxGap}{Maximum distance between CpGs to be included in the same region. Default is 1000 bp}
}
\value{
Data.frame of background regions with location, number of CpGs, and width
}
\description{
Get background regions from filtered bsseq object based on minCpGs and maxGap
}
