% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{getChrom}
\alias{getChrom}
\title{getChrom}
\usage{
getChrom(bsseq = bs.filtered.bsseq)
}
\arguments{
\item{bsseq}{Smoothed bsseq object}
}
\value{
Tibble of smoothed chromosomal methylation values and phentoype data
}
\description{
Gets smoothed chromosomal methylation values
}
