% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{getCpGs}
\alias{getCpGs}
\title{Obtain CpG island, CpG shore, CpG shelf, and open sea annotations}
\usage{
getCpGs(genome = genome, resPath)
}
\arguments{
\item{genome}{Character specifying the genome}

\item{resPath}{character specifying path to local resources in internet is not available}
}
\value{
A \code{GRanges} object of CpG island, CpG shore, CpG shelf, and open sea annotations
}
\description{
Obtain UCSC CpG islands and build CpG shore, CpG shelf, and open sea annotations.
 This function is based on \code{annotatr:::build_cpg_annots()}; however, 
 it obtains annotations for all genomes in the UCSC genome browser.
}
\details{
getCpGs
}
\references{
Based on \code{annotatr:::build_cpg_annots()},
 see: \url{https://github.com/rcavalcante/annotatr/blob/master/R/build_annotations.R}
}
