% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{getGenes}
\alias{getGenes}
\title{Obtain CpG island, CpG shore, CpG shelf, and open sea annotations}
\usage{
getGenes(TxDb = TxDb, annoDb = annoDb)
}
\arguments{
\item{TxDb}{A TxDb or EnsDb annotation object}

\item{annoDb}{Character string naming the OrgDb package, or an OrgDb object}

\item{genome}{Character specifying the genome}
}
\value{
A \code{GRanges} object of CpG island, CpG shore, CpG shelf, and open sea annotations

A GRanges object of annotated exons for plotting
}
\description{
Obtain UCSC CpG islands and build CpG shore, CpG shelf, and open sea annotations.
 This function is based on \code{annotatr:::build_cpg_annots()}; however, 
 it obtains annotations for all genomes in the UCSC genome browser.

Obtain exon/CDS annotations and format for plotDMRs()
}
\details{
getCpGs
}
\references{
Based on \code{annotatr:::build_cpg_annots()},
 see: \url{https://github.com/rcavalcante/annotatr/blob/master/R/build_annotations.R}
}
