% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{getGlobal}
\alias{getGlobal}
\title{getGlobal}
\usage{
getGlobal(bsseq = bs.filtered.bsseq)
}
\arguments{
\item{bsseq}{Smoothed bsseq object}
}
\value{
Tibble of smoothed global methylation values and phentoype data
}
\description{
Gets smoothed global methylation values
}
