% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getSmooth}
\alias{getSmooth}
\title{getSmooth}
\usage{
getSmooth(bsseq = bs.filtered.bsseq, regions = sigRegions)
}
\arguments{
\item{bsseq}{A smoothed \code{bsseq} object}

\item{regions}{A \code{GRanges} object of regions to obtain smoothed methylation values for}
}
\value{
A data frame of individual smoothed methylation values
}
\description{
Provides individual smoothed methylation values for a \code{GRanges} object using \code{bsseq}
}
