% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalPlots.R
\name{globalPlots}
\alias{globalPlots}
\title{globalPlots}
\usage{
globalPlots(plotMatrix, bs.filtered.bsseq = bs.filtered.bsseq)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}

\item{matrix}{A matrix of smoothed individual methylation values to plot}
}
\value{
Saves PCA, density and interactive MDS plots
}
\description{
Creates PCA, density, and interactive MDS plots of global methylation data
}
