% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalStats.R
\name{globalStats}
\alias{globalStats}
\title{globalStats}
\usage{
globalStats(
  bs.filtered.bsseq = bs.filtered.bsseq,
  genome = genome,
  testCovariate = testCovariate,
  adjustCovariate = NULL,
  matchCovariate = NULL
)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with design matrix in pData}

\item{genome}{Character specifying the genome of interest}

\item{testCovariate}{The factor to test for differences between groups}

\item{adjustCovariate}{The covariate(s) to adjust for between groups}

\item{matchCovariate}{Another covariate to adjust for between groups (for dmrseq compatibility)}
}
\value{
A list of tibbles with smoothed global and chromosomal methylation statistics
 and the values used for the tests
}
\description{
Computes the average smoothed global and chromosomal CpG methylation values
 for each sample and tests for differences between groups while adjusting for the provided 
 covariates. CpG island testing is performed for the human, mouse, and rate genomes.
 Global methylation and CpG island differences are tested for using an ANOVA through the 
 \code{\link[stats]{aov}} function. The chromosomal methylation differences are tested using 
 pairwise comparisons calculated from contrasts of the factor of interest via 
 the \code{\link[lsmeans]{lsmeans}} package.
}
\references{
\url{https://cran.r-project.org/web/packages/broom/vignettes/broom_and_dplyr.html}

\url{https://www.jstatsoft.org/article/view/v069i01/v69i01.pdf}
}
