% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{GO}
\alias{GO}
\title{GO}
\usage{
GO(
  sigRegions = sigRegions,
  regions = regions,
  annoDb = annoDb,
  TxDb = TxDb,
  genome = genome,
  GOfuncR = TRUE
)
}
\arguments{
\item{sigRegions}{\code{GRanges} object of DMRs.}

\item{regions}{\code{GRanges} object of background regions.}

\item{annoDb}{Character specifying \code{OrgDb} annotation package for species of interest.}

\item{TxDb}{\code{TxDb} or \code{EnsDb} annotation package for genome of interest.}

\item{genome}{Character specifying gnome.}

\item{GOfuncR}{Logical indicating whether a time consuming GOfuncR analysis should be run.}
}
\value{
Spreadsheets and plots of the GO enrichment results.
}
\description{
Perform Gene Ontology enrichment analysis of DMRs from \code{enrichR}, \code{rGREAT}, 
and \code{GOfuncR},slims the terms using \code{rrvgo}, and plots the results.
}
