% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOplot.R
\name{GOplot}
\alias{GOplot}
\title{GOplot}
\usage{
GOplot(GO = GO)
}
\arguments{
\item{GO}{A list of ontology and pathway data frames returned from \code{enrichR::enrichr()}}
}
\value{
A \code{ggplot} object of top significant GO and KEGG pathway terms from an enrichR analysis
 that can be viewed by calling it, saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax
}
\description{
Plots top signficant Gene Ontology and KEGG pathway terms from enrichR
}
