% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{gr2bed}
\alias{gr2bed}
\title{gr2bed}
\usage{
gr2bed(gr = gr, bed = bed)
}
\arguments{
\item{gr}{Genomic ranges or bsseq object}

\item{bed}{Name of the bed file in quotations}
}
\value{
Bed file
}
\description{
Save a genomic ranges object as a basic bed file
}
