% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{gr2csv}
\alias{gr2csv}
\title{gr2csv}
\usage{
gr2csv(gr = gr, csv = csv)
}
\arguments{
\item{gr}{Genomic ranges or bsseq object}

\item{csv}{Name of the csv file in quotations}
}
\value{
CSV file
}
\description{
Save a genomic ranges object as a csv file
}
