% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREAT.R
\name{GREAT}
\alias{GREAT}
\title{GREAT}
\usage{
GREAT(sigRegions = sigRegions, regions = regions, genome = genome)
}
\arguments{
\item{regions}{A \code{GRanges} object of background regions returned by \code{dmrseq:dmrseq()}}

\item{genome}{A character vector specifying the genome of interest ("hg38" or "mm10")}

\item{siRegions}{A \code{GRanges} object of signficant DMRs returned by \code{dmrseq:dmrseq()}}
}
\value{
A list of gene ontology enrichments.
}
\description{
Note: Depracted by GREAT version 4.0 update. Perform gene ontology of DMRs relative to background regions from \code{dmrseq::dmrseq()} using \code{rGREAT}.
}
