% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS_specific.R
\name{GREAT_analysis}
\alias{GREAT_analysis}
\title{GREAT analysis}
\usage{
GREAT_analysis(
  gr,
  genesetName = "GO:BP",
  padj_cutoff = 0.2,
  status = "hyper",
  dname = "sigDMR",
  geneset_cutoff = 200,
  dir = "./",
  genome = "hg38"
)
}
\arguments{
\item{gr}{a GRanges object.}

\item{padj_cutoff}{Numeric cutoff value [from 0 to 1] for the adjusted p-value.}

\item{status}{indicates the direction of methylation changes, either hyper or hypo.}

\item{dname}{indicates nature of DMR, either sigDMR or background}

\item{geneset_cutoff}{cutoff value for the maximal size of the geneset.}

\item{dir}{path for saving GREAT analysis results.}

\item{genome}{Character specifying the genome.}
}
\description{
Perform GREAT analysis for DMR
}
\details{
GREAT_analysis
}
