% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{Houseman}
\alias{Houseman}
\title{Houseman method for blood cell composition estimation}
\usage{
Houseman(bs.filtered.bsseq.cc = bs.filtered.bsseq.cc)
}
\arguments{
\item{bs.filtered.bsseq.cc}{A \code{bsseq} object that has been lifted over to hg19}
}
\value{
Cell composition estimates
}
\description{
Utilize the Houseman method to estimate whole-blood cell composition
}
\details{
Houseman
}
