% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprintOverlap.R
\name{imprintOverlap}
\alias{imprintOverlap}
\title{imprintOverlap}
\usage{
imprintOverlap(
  sigRegions = sigRegions,
  regions = regions,
  TxDb = TxDb,
  annoDb = annoDb
)
}
\arguments{
\item{sigRegions}{\code{GRanges} object of DMRs.}

\item{regions}{\code{GRanges} object of background regions.}

\item{TxDb}{TxDb annotation package for genome of interest.}

\item{annoDb}{Character specifying OrgDb annotation package for species of interest.}
}
\value{
A printed statement with the enrichment results.
}
\description{
Test DMRs for overlaps with human imprinted genes from
 \url{https://www.geneimprint.com/site/genes-by-species.Homo+sapiens.imprinted-All}
}
