% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{labelDirection}
\alias{labelDirection}
\title{labelDirection}
\usage{
labelDirection(regions = sigRegions)
}
\arguments{
\item{regions}{A \code{GRanges} object returned by \code{dmrseq:dmrseq()}.}
}
\value{
An \code{GRanges} object annotated with directionality of change.
}
\description{
Annotate a \code{GRanges} object containing either DMRs or background regions with directionality of change
 (i.e. hypermethlated or hypomethylated).
}
