% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{lift450k}
\alias{lift450k}
\title{lift450k}
\usage{
lift450k(probes = probes)
}
\arguments{
\item{probes}{A dataframe or vector of 450k CpG IDs}
}
\value{
A \code{GRanges} object of hg38 coordinates
}
\description{
LiftOver 450k CpG IDs to hg38 coordinates
}
