% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{manQQ}
\alias{manQQ}
\title{manQQ}
\usage{
manQQ(backgroundAnno = backgroundAnno, ...)
}
\arguments{
\item{...}{Additional arguments passed onto \code{\link[CMplot]{CMplot}}}

\item{peakAnno}{A \code{ChIPseeker csAnno} peak object of background regions from \code{dmrseq::dmrseq()}}
}
\value{
Saves a pdf of manhattan and qq plots
}
\description{
Create manhattan and Quantile-Quantile (Q-Q) plots of \code{ChIPseeker csAnno} peak object with genic annotations using \code{CMplot}
}
