% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylLearn.R
\name{methylLearn}
\alias{methylLearn}
\title{methylLearn}
\usage{
methylLearn(bsseq = bs.filtered.bsseq, regions = sigRegions,
  testCovariate = testCovariate, TxDb = NA, annoDb = NA)
}
\arguments{
\item{bsseq}{Smoothed bsseq object.}

\item{regions}{Genomic ranges object.}

\item{testCovariate}{Factor of interest.}

\item{TxDb}{TxDb annotation package for genome of interest.}

\item{annoDb}{Character specifying OrgDb annotation package for species of interest.}
}
\value{
A tibble object containing the variable importance ranks and annotations of the most important DMRs, which overlap between the 
top 1% of DMRs in the two DMR ranking lists obtained from feature selection methods using RF and SVM algorithsms.
}
\description{
Performs feature selection on significant DMRs (predictors) based on random forest (RF) and support vector machine (SVM)
algorithms to generate two lists of DMRs ranked by order of importance. Then finds and annotates the DMRs that overlap between the 
top 1% of DMRs in the two DMR ranking lists.
}
\references{
\url{https://www.analyticsvidhya.com/blog/2016/03/select-important-variables-boruta-package/}
}
