% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS_specific.R
\name{output_DMR}
\alias{output_DMR}
\title{Output DMRs}
\usage{
output_DMR(DMR)
}
\arguments{
\item{DMR}{a list of two dataframes.}

\item{dir}{path for saving DMR bed files.}
}
\description{
Write DMR to bed files
}
\details{
output_DMR
}
