% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{packageLoad}
\alias{packageLoad}
\title{packageLoad}
\usage{
packageLoad(packages = packages)
}
\arguments{
\item{packages}{Character string or desired packages}
}
\description{
Install and load desired packages
}
