% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{prepareCC}
\alias{prepareCC}
\title{prepareCC}
\usage{
prepareCC(bs.filtered.bsseq = bs.filtered.bsseq, genome = genome)
}
\arguments{
\item{bs.filtered.bsseq}{A \code{bsseq} object}

\item{genome}{Character string of genome symbol c("hg38", "hg19").}
}
\value{
A \code{bsseq} object with hg19 coordinates
}
\description{
Prepare a \code{bsseq} object for cell composition estimation
}
