% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{read_excel_all}
\alias{read_excel_all}
\title{read_excel_all}
\usage{
read_excel_all(filename)
}
\arguments{
\item{filename}{A character vector specifying the name of the excel document}
}
\value{
A list of \code{tibbles} with the excel document data
}
\description{
Read all sheets in an excel document
}
\references{
\url{https://stackoverflow.com/a/12945838}
}
