% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{REVIGO}
\alias{REVIGO}
\title{REVIGO}
\usage{
REVIGO(GO = GO, tool = c("enrichR", "rGREAT", "GOfuncR"))
}
\arguments{
\item{GO}{A dataframe or list of dataframes returned
from \code{enrichR::enrichr()}, \code{rGREAT::getEnrichmentTables()}, or \code{GOfuncR::go_enrich()}.}

\item{tool}{A character vector of the name of the database (enrichR, rGREAT, or GOfuncR).}
}
\value{
A \code{tibble} of top distinct and significant GO terms from an \code{enrichR} 
or \code{rGREAT} analysis.
}
\description{
Slims top signficant Gene Ontology terms from enrichR, rGREAT, and GOfuncR.
The terms are ranked by dispensability.
}
\references{
\url{https://github.com/hbc/revigoR/blob/master/rvest_revigo.R}

\url{http://revigo.irb.hr}
}
