% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOLA.R
\name{roadmap}
\alias{roadmap}
\title{roadmap}
\usage{
roadmap(sigRegions = sigRegions, regions = regions, cores = cores)
}
\arguments{
\item{sigRegions}{A \code{GRanges} object of significant regions}

\item{regions}{A \code{GRanges} object of background regions}

\item{cores}{An integer of how many cores to use}
}
\value{
A \code{tibble} of enrichment results
}
\description{
Perfom enrichment testing against the Roadmap Epigenomics core histone modifications
 (5 marks, 127 epigenomes) for hg38 using \code{LOLA}
}
