% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOLA.R
\name{roadmap_heatmap}
\alias{roadmap_heatmap}
\title{roadmap_heatmap}
\usage{
roadmap_heatmap(roadmap = roadmap)
}
\arguments{
\item{chromHMM}{A \code{tibble} of enrichment results}
}
\value{
Saves a heatmap
}
\description{
Plot a heatmap of \code{LOLA} enrichment testing results of the Roadmap Epigenomics
 core marks for hg38
}
