% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{saveExternal}
\alias{saveExternal}
\title{Save regions for external enrichment testing}
\usage{
saveExternal(sigRegions = sigRegions, regions = regions)
}
\arguments{
\item{sigRegions}{A \code{GRanges} object of signficant DMRs returned by \code{dmrseq::dmrseq()}}

\item{regions}{A \code{GRanges} object of background regions returned by \code{dmrseq::dmrseq()}}
}
\value{
Saves external GAT and HOMER folders with bed files into the working directory
}
\description{
Save DMRs and background regions from \code{dmrseq::dmrseq()} in formats for external analyses using GAT and HOMER
}
\details{
saveExternal
}
