% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{singleCpGPCA}
\alias{singleCpGPCA}
\title{singleCpGPCA}
\usage{
singleCpGPCA(bs.filtered.bsseq = bs.filtered.bsseq, group = NA)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}

\item{group}{Ordered character vector of sample groupings used to assign colors}
}
\value{
A \code{ggplot} object that can be viewed by calling it,
 saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Performs and plots a PCA of single CpG individual smoothed methylation values
}
\references{
\url{https://stackoverflow.com/questions/40315227/how-to-solve-prcomp-default-cannot-rescale-a-constant-zero-column-to-unit-var/40317343}
}
