% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{slimGO}
\alias{slimGO}
\title{Slim GO results}
\usage{
slimGO(
  GO = GO,
  tool = c("enrichR", "rGREAT", "GOfuncR"),
  annoDb = annoDb,
  plots = FALSE,
  threshold = 0.7
)
}
\arguments{
\item{GO}{A dataframe or list of dataframes returned from \code{enrichR::enrichr()}, 
\code{rGREAT::getEnrichmentTables()},or \code{GOfuncR::go_enrich()}.}

\item{tool}{A character vector of the name of the database (enrichR, rGREAT, or GOfuncR).}

\item{annoDb}{Character specifying \code{OrgDb} annotation package for species of interest.}

\item{plots}{Logical indicating if scatter and treemap plots should be generated.}

\item{threshold}{Numeric indicating similarity threshold (0-1) for \code{rrvgo::reduceSimMatrix()}.
0.9 is large, 0.7 is medium, 0.5 is small, and 0.4 is tiny. Default is 0.7.}
}
\value{
A \code{tibble} of top distinct and significant GO terms from an \code{enrichR},
 \code{rGREAT} or \code{GOfuncR} analysis.
}
\description{
Slims top significant Gene Ontology terms from \code{enrichR}, \code{rGREAT}, 
and \code{GOfuncR} using \code{rrvgo}.
}
\details{
slimGO
}
