% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothANOVA.R
\name{smoothANOVA}
\alias{smoothANOVA}
\title{smoothANOVA}
\usage{
smoothANOVA(smoothAvg = smoothAvg)
}
\arguments{
\item{smoothAvg}{Tibble of average smoothed methylation values and covariates}
}
\value{
Excel spreadsheet summariazing the ANOVA(s), where a clean statistical structure means no random effects, multiple testing corrections, or type III ANOVAs are needed
}
\description{
Perform an ANOVA on smoothed methylation values averaged globally or across chromosomes
}
\references{
\url{https://cran.r-project.org/web/packages/broom/vignettes/broom_and_dplyr.html}
}
