% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothHeatmap.R
\name{smoothHeatmap}
\alias{smoothHeatmap}
\title{smoothHeatmap}
\usage{
smoothHeatmap(regions = sigRegions, bsseq = bs.filtered.bsseq,
  groups = bs.filtered.bsseq \%>\% pData() \%>\% as.tibble() \%>\%
  pull(!!testCovariate),
  out = "sig_individual_smoothed_DMR_methylation.txt", ...)
}
\arguments{
\item{regions}{Genomic ranges object}

\item{bsseq}{Smoothed bsseq object}

\item{groups}{Ordered test covariate information for each sample}

\item{out}{Name of the text file to save in quotations}

\item{...}{Additional arguments passed onto heatmap.3()}

\item{names}{Ordered sample names}
}
\value{
Saves a pdf image of the heatmap
}
\description{
Plot a heatmap of normalized individual smoothed methylation value z scores for selected regions (i.e. significant DMRs)
}
\references{
\url{https://sebastianraschka.com/Articles/heatmaps_in_r.html}

\url{https://raw.githubusercontent.com/obigriffith/biostar-tutorials/master/Heatmaps/heatmap.3.R}

\url{https://www.biostars.org/p/18211/}
}
