% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothHeatmap.R
\name{smoothPheatmap}
\alias{smoothPheatmap}
\title{DMR heatmap}
\usage{
smoothPheatmap(
  bs.filtered.bsseq = bs.filtered.bsseq,
  sigRegions = sigRegions,
  testCovariate = testCovariate,
  ...
)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object}

\item{sigRegions}{\code{GRanges} object of regions to plot a heatmap for}

\item{testCovariate}{The factor tested for differences between groups}

\item{...}{Additional arguments passed onto \code{pheatmap()}}
}
\value{
Saves a pdf image of the heatmap in the DMR folder
}
\description{
Plot a heatmap of normalized individual smoothed methylation value z scores for selected regions (i.e. significant DMRs)
}
\details{
smoothPheatmap
}
\references{
\url{https://davetang.org/muse/2018/05/15/making-a-heatmap-in-r-with-the-pheatmap-package/}
}
