% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tidyDMRs}
\alias{tidyDMRs}
\title{tidyDMRs}
\usage{
tidyDMRs(regions = peakAnno)
}
\arguments{
\item{regions}{Peak file or GRanges object from ChIPseeker}
}
\description{
Tidy DMRs or background regions that have been annotated using ChIPseeker
}
