% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{windows}
\alias{windows}
\title{windows}
\usage{
windows(bs.filtered.bsseq = bs.filtered.bsseq, size = 20000, goi = goi)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}

\item{size}{The number of bases in the window (default is 2e4, which is 20 Kb)}

\item{goi}{A \code{BSgenome} object of the genome of interest (i.e. "BSgenome.Hsapiens.UCSC.hg38")}
}
\value{
A matrix of smoothed individual methylation values
}
\description{
Obtain windows of individual smoothed methylation values
}
