% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{windowsPCA}
\alias{windowsPCA}
\title{windowsPCA}
\usage{
windowsPCA(bsseq = bs.filtered.bsseq, goi = goi)
}
\arguments{
\item{bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}

\item{goi}{A \code{BSgenome} object of the genome of interest (i.e. "BSgenome.Hsapiens.UCSC.hg38")}
}
\value{
A \code{ggplot} object that can be viewed by calling it,
 saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Performs and plots a PCA of 20kb windows from individual smoothed methylation values
}
