% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{arrayRanges}
\alias{arrayRanges}
\title{Obtain EPIC array coordinates}
\usage{
arrayRanges()
}
\value{
A \code{GRanges} object of hg19 EPIC coordinates
}
\description{
Obtain hg19 EPIC array coordinates
}
\details{
arrayRanges
}
\references{
\url{https://support.bioconductor.org/p/78652/}
}
