% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{bsseqLift}
\alias{bsseqLift}
\title{bsseqLift}
\usage{
bsseqLift(bs.filtered.bsseq = bs.filtered.bsseq)
}
\arguments{
\item{bs.filtered.bsseq}{A \code{bsseq} object with hg38 coordinates}
}
\value{
A \code{bsseq} object with hg19 coordinates
}
\description{
LiftOver a hg38 bsseq objet to hg19 coordinates
}
