% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalPlots.R
\name{CpGs}
\alias{CpGs}
\title{Extract single CpG methylation values}
\usage{
CpGs(bs.filtered.bsseq = bs.filtered.bsseq)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}
}
\value{
A matrix of smoothed individual methylation values
}
\description{
Extracts single CpG individual smoothed methylation values
}
\details{
CpGs
}
