% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{densityPlot}
\alias{densityPlot}
\title{densityPlot}
\usage{
densityPlot(matrix = matrix, group = NA)
}
\arguments{
\item{matrix}{A matrix of smoothed individual methylation values}

\item{group}{Ordered factor vector of sample groupings}
}
\value{
A \code{ggplot} object that can be viewed by calling it,
 saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Creates a density plot of the mean of individual smoothed methylation values
}
