% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{densityPlot}
\alias{densityPlot}
\title{densityPlot}
\usage{
densityPlot(bs.filtered.bsseq = bs.filtered.bsseq, group = NA)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with a testCovariate in \code{pData}}

\item{group}{Ordered character vector of sample groupings used to assign colors}
}
\value{
A \code{ggplot} object that can be viewed by calling it,
 saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Creates a density plot of the mean of individual smoothed methylation values for every CpG
}
