% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRich.R
\name{DMparseR}
\alias{DMparseR}
\title{DMparseR}
\usage{
DMparseR(
  direction = c("All DMRs", "Hypermethylated DMRs", "Hypomethylated DMRs"),
  type = c("CpG", "genic")
)
}
\arguments{
\item{direction}{A character vector of the DMR profiles to analyze
c("All DMRs", "Hypermethylated DMRs", "Hypomethylated DMRs").}

\item{type}{A character vector of the type of results to parse i.e. c("CpG", "genic").}
}
\value{
A \code{tibble} of enrichment results
}
\description{
Parse the results from multiple \code{DMRichR} runs.
}
