% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{DMReport}
\alias{DMReport}
\title{DMReport}
\usage{
DMReport(
  sigRegions = sigRegions,
  regions = regions,
  bsseq = bs.filtered.bsseq,
  coverage = coverage,
  name = "DMReport"
)
}
\arguments{
\item{sigRegions}{\code{GRanges} object of signficant regions (DMRs or blocks) from \code{dmrseq} that 
were annotated by \code{DMRichR::annotateRegions}}

\item{regions}{\code{GRanges} object of background regions from \code{dmrseq}}

\item{bsseq}{Smoothed \code{bsseq} object}

\item{coverage}{Numeric of coverage samples were filtered for}

\item{name}{Character for html report name}
}
\value{
Saves an html report of DMRs with genic annotations
}
\description{
Create an html report of significant regions from \code{dmrseq}
}
