% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{DMReport}
\alias{DMReport}
\title{DMReport}
\usage{
DMReport(sigRegionsAnno = sigRegionsAnno)
}
\arguments{
\item{sigRegionsAnno}{A \code{ChIPseeker csAnno} peak object of DMRs from \code{dmrseq::dmrseq()}.}
}
\value{
Saves an html report of DMRs with genic annotations.
}
\description{
Create an html report of a \code{ChIPseeker csAnno} peak object with genic annotations.
}
