% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRich.R
\name{DMRichCpG}
\alias{DMRichCpG}
\title{DMRichCpG}
\usage{
DMRichCpG(sigRegions = sigRegions, regions = regions, genome = genome)
}
\arguments{
\item{sigRegions}{\code{GRanges} object of DMRs}

\item{regions}{\code{GRanges} object of background regions}

\item{genome}{A character vector specifying the genome of interest
c("hg38", "hg19", "mm10", "mm9", "rn6", "rn5")}
}
\value{
A tibble with the enrichment results
}
\description{
Test DMRs for overlaps with CpG annotations for human, mouse, or rat
}
