% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{getExons}
\alias{getExons}
\title{getExons}
\usage{
getExons(TxDb = TxDb)
}
\arguments{
\item{TxDb}{A \code{ensDb} object}
}
\value{
A \code{GRanges object} of annotated exons for every gene with a symbol in the genome.
}
\description{
Obtain exon annotations from a \code{ensDb} object and format for \code{plotDMRs()}
}
\references{
Based on \code{annotatr::build_gene_annots()},
 see: \url{https://github.com/rcavalcante/annotatr/blob/master/R/build_annotations.R}
}
