% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{getGenes}
\alias{getGenes}
\title{Obtain exon annotations for plotting from a TxDb or EnsDb}
\usage{
getGenes(TxDb = TxDb, annoDb = annoDb)
}
\arguments{
\item{TxDb}{A TxDb or EnsDb annotation object}

\item{annoDb}{Character string naming the OrgDb package, or an OrgDb object}
}
\value{
A GRanges object of annotated exons for plotting
}
\description{
Obtain exon/CDS annotations and format for plotDMRs()
}
\details{
getGenes
}
