% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getSmooth}
\alias{getSmooth}
\title{getSmooth}
\usage{
getSmooth(bsseq = bsseq, regions = regions, out = out)
}
\arguments{
\item{bsseq}{Smoothed bsseq object}

\item{regions}{Genomic ranges object}

\item{out}{Name of the text file in quotations}
}
\value{
Genomic ranges object of individual smoothed methylation values and text file
}
\description{
Provides individual smoothed methylation values for genomic ranges objects using bsseq
}
