% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{ggplot2 colors}
\usage{
gg_color_hue(n = n)
}
\arguments{
\item{n}{Number of samples}
}
\value{
Character string of colors
}
\description{
Generate ggplot2 style colors
}
\details{
gg_color_hue
}
\references{
\url{https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
}
