% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalStats.R
\name{globalStats}
\alias{globalStats}
\title{globalStats}
\usage{
globalStats(bsseq = bs.filtered.bsseq, test = testCovariate,
  adjust = adjustCovariate, match = matchCovariate)
}
\arguments{
\item{bsseq}{Smoothed bsseq object with design matrix in pData}

\item{test}{The factor to test for differences between groups}

\item{adjust}{The covariate(s) to adjust for between groups}

\item{match}{Another covariate to adjust for between groups (for dmrseq compatibility)}
}
\value{
A list of tibbles with smoothed global and chromosomal methylation statsitics and values
}
\description{
Obtains smoothed global and chromosomal methylation values and tests for differences between groups while adjusting for the provided covariates
}
\references{
\url{https://cran.r-project.org/web/packages/broom/vignettes/broom_and_dplyr.html}
}
