% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalStats.R
\name{globalStats}
\alias{globalStats}
\title{Test for global methylation differences}
\usage{
globalStats(
  bs.filtered.bsseq = bs.filtered.bsseq,
  genome = genome,
  testCovariate = testCovariate,
  adjustCovariate = NULL,
  matchCovariate = NULL
)
}
\arguments{
\item{bs.filtered.bsseq}{Smoothed \code{bsseq} object with design matrix in pData}

\item{genome}{Character specifying the genome of interest}

\item{testCovariate}{The factor to test for differences between groups}

\item{adjustCovariate}{The covariate(s) to adjust for between groups}

\item{matchCovariate}{Another covariate to adjust for between groups (for dmrseq compatibility)}
}
\value{
A list of tibbles with smoothed global and CpG island methylation statistics
 and the values used for the tests
}
\description{
Computes the average smoothed global CpG methylation values
 for each sample and tests for differences between groups while adjusting for the provided 
 covariates. CpG island testing is performed for the human, mouse, and rat genomes.
 Global methylation and CpG island differences are tested for using an ANOVA through the 
 \code{\link[stats]{aov}} function.
}
\details{
globalStats
}
