% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{GOplot}
\alias{GOplot}
\title{GOplot}
\usage{
GOplot(revigoResults = revigoResults)
}
\arguments{
\item{revigoResults}{A \code{tibble} from\code{DMRichR::REVIGO()}.}
}
\value{
A \code{ggplot} object of top significant GO and pathway terms from an \code{enrichR} 
or \code{rGREAT} analysis that can be viewed by calling it, saved with \code{ggplot2::ggsave()}, 
or further modified by adding \code{ggplot2} syntax.
}
\description{
Slims and plots top significant Gene Ontology terms from enrichR, rGREAT, and GOfuncR.
The terms are ranked by dispensability before being plotted, which then orders them by p-value.
}
