% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{gr2csv}
\alias{gr2csv}
\title{gr2csv}
\usage{
gr2csv(gr = gr, csv = csv)
}
\arguments{
\item{gr}{\code{GRanges} or \code{bsseq} object}

\item{csv}{Character string of save file name}
}
\value{
Saves a CSV file
}
\description{
Save a genomic ranges object as a csv file
}
