% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{Manhattan}
\alias{Manhattan}
\title{Manhattan plot}
\usage{
Manhattan(backgroundAnno = backgroundAnno, ...)
}
\arguments{
\item{backgroundAnno}{A \code{tibble} of annotated background regions from \code{dmrseq::dmrseq()}}

\item{...}{Additional arguments passed onto \code{\link[CMplot]{CMplot}}}
}
\value{
Saves a pdf of manhattan and qq plots
}
\description{
Create a Manhattan plot of \code{ChIPseeker csAnno} peak object with genic annotations using \code{CMplot}
}
\details{
Manhattan
}
