% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylLearn.R
\name{methylLearn}
\alias{methylLearn}
\title{methylLearn}
\usage{
methylLearn(data = getMeth(BSseq = bs.filtered.bsseq, regions =
  sigRegions, type = "smooth", what = "perRegion") \%>\% as.matrix() \%>\%
  t(), groups = bs.filtered.bsseq \%>\% pData() \%>\% as_tibble() \%>\%
  pull(!!testCovariate), k.fold = 10, repeats = 3, ...)
}
\arguments{
\item{groups}{Ordered test covariate information for each sample}

\item{k.fold}{number of folds in the cross-validation}

\item{repeats}{number of repeats in the cross-validation}

\item{...}{Additional arguments passed onto randomForest::randomForest()}

\item{bsseq}{Smoothed bsseq object}

\item{regions}{Genomic ranges object}

\item{names}{Ordered sample names}
}
\value{
?
}
\description{
WORK IN PROGRESS ... Performs random forest machine learning with repeated k-fold cross validation on significant DMRs to identify the most important DMRs
}
\references{
\url{https://cran.rstudio.com/web/packages/randomForestExplainer/vignettes/randomForestExplainer.html}

\url{https://stats.stackexchange.com/questions/82162/cohens-kappa-in-plain-english}

\url{https://explained.ai/rf-importance/}

\url{https://machinelearningmastery.com/feature-selection-with-the-caret-r-package/}

\url{http://ml-tutorials.kyrcha.info/rfe.html}
}
