% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA}
\usage{
PCA(matrix = matrix, group = bs.filtered.bsseq \%>\% pData() \%>\%
  dplyr::as_tibble() \%>\% dplyr::pull(!!testCovariate), title = title)
}
\arguments{
\item{matrix}{Matrix of transposed individual methylation values}

\item{title}{Character string of title for plot and pdf}
}
\value{
A \code{ggplot} object that can be viewed by calling it,
 saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Performs and plots a PCA from individual smoothed methylation values
}
