% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA}
\usage{
PCA(matrix = matrix, group = NA)
}
\arguments{
\item{matrix}{A matrix of smoothed individual methylation values}

\item{group}{Ordered factor vector of sample groupings}
}
\value{
A \code{ggplot} object that can be viewed by calling it,
 saved with \code{ggplot2::ggsave()}, or further modified by adding \code{ggplot2} syntax.
}
\description{
Performs and plots a PCA from individual smoothed methylation values
}
\references{
\url{https://stackoverflow.com/questions/40315227/how-to-solve-prcomp-default-cannot-rescale-a-constant-zero-column-to-unit-var/40317343}
}
