% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA}
\usage{
PCA(matrix = matrix, title = title)
}
\arguments{
\item{matrix}{Matrix of transposed individual methylation values}

\item{title}{Character string of title for plot and pdf}
}
\value{
PCA plot
}
\description{
Provides individual smoothed methylation values for genomic ranges objects using bsseq
}
