% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{prepareHOMER}
\alias{prepareHOMER}
\title{Save regions for HOMER}
\usage{
prepareHOMER(sigRegions = sigRegions, regions = regions)
}
\arguments{
\item{sigRegions}{A \code{GRanges} object of significant DMRs returned by \code{dmrseq::dmrseq()}}

\item{regions}{A \code{GRanges} object of background regions returned by \code{dmrseq::dmrseq()}}
}
\value{
Creates a folder for HOMER with bed files
}
\description{
Save DMRs and background regions from \code{dmrseq::dmrseq()}for HOMER
}
\details{
prepareHOMER
}
