% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBismark.R
\name{processBismark}
\alias{processBismark}
\title{processBismark}
\usage{
processBismark(files = list.files(path = getwd(), pattern = "*.txt.gz"),
  names = gsub("_.*$", "", list.files(path = getwd(), pattern =
  "*.txt.gz")), meta = read.csv("sample_info.csv", header = TRUE),
  groups = testCovariate, Cov = coverage, mc.cores = cores)
}
\arguments{
\item{files}{List of cytosine report file paths}

\item{names}{Ordered character vector of sample names}

\item{meta}{Design matrix table with sample name in the Name column}

\item{groups}{Factor of interest (testCovariate)}

\item{mc.cores}{Number of cores to use}
}
\description{
Process bismark cytosine reports into bsseq objects with design matrix pData
}
