% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBismark.R
\name{processBismark}
\alias{processBismark}
\title{processBismark}
\usage{
processBismark(
  files = list.files(path = getwd(), pattern = "*.txt.gz"),
  meta = openxlsx::read.xlsx("sample_info.xlsx", colNames = TRUE) \%>\%
    dplyr::mutate_if(is.character, as.factor),
  testCovar = testCovariate,
  adjustCovar = NULL,
  matchCovar = NULL,
  Cov = coverage,
  mc.cores = cores,
  per.Group = perGroup,
  sexCheck = FALSE
)
}
\arguments{
\item{files}{List of cytosine report file paths}

\item{meta}{Design matrix data frame with sample name in the Name column}

\item{testCovar}{Factor of interest (testCovariate)}

\item{adjustCovar}{Variables to adjust for (adjustCovariate)}

\item{matchCovar}{Variable to block for when constructing permutations (matchCovariate)}

\item{Cov}{CpG coverage cutoff (1x recommended)}

\item{mc.cores}{Integer specifying the number of cores to use}

\item{per.Group}{Percent of samples per a group to apply the CpG coverage cutoff to (from 0 to 1)}

\item{sexCheck}{Logical (TRUE or FALSE) indicating whether to confirm the sex of samples.
This requires a column called "Sex" (case sensitive) in sample_info.xlsx. 
Males should be coded as either "Male", "male", "M", or "m".
Females coded as "Female", "female", "F", or "f".}
}
\description{
Process bismark cytosine reports into bsseq objects with design matrix pData
}
