% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS_specific.R
\name{processReport}
\alias{processReport}
\title{Process methylation report files}
\usage{
processReport(design, cores)
}
\arguments{
\item{design}{sample information dataframe.}

\item{cores}{Numeric specifying the number of cores to use. 10 is recommended.}
}
\value{
a bsseq object
}
\description{
Build bsseq object from report files and perform smoothing
}
\details{
processReport
}
