% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{slimGO}
\alias{slimGO}
\title{slimGO}
\usage{
slimGO(GO = GO, tool = c("enrichR", "rGREAT", "GOfuncR"), annoDb = annoDb)
}
\arguments{
\item{GO}{A dataframe or list of dataframes returned from \code{enrichR::enrichr()}, 
\code{rGREAT::getEnrichmentTables()},or \code{GOfuncR::go_enrich()}.}

\item{tool}{A character vector of the name of the database (enrichR, rGREAT, or GOfuncR).}

\item{annoDb}{Character specifying \code{OrgDb} annotation package for species of interest.}
}
\value{
A \code{tibble} of top distinct and significant GO terms from an \code{enrichR},
 \code{rGREAT} or \code{GOfuncR} analysis.
}
\description{
Slims top significant Gene Ontology terms from enrichR, rGREAT, and GOfuncR 
using \code{rrvgo}.
}
