% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{smooth2txt}
\alias{smooth2txt}
\title{Save regions and methylation values}
\usage{
smooth2txt(bsseq = bs.filtered.bsseq, regions = sigRegions, txt = txt)
}
\arguments{
\item{bsseq}{A smoothed \code{bsseq} object}

\item{regions}{A \code{GRanges} object of regions to obtain smoothed methylation values for}

\item{txt}{Character string of save file name}
}
\value{
Saves a text file
}
\description{
Provides individual smoothed methylation values for a \code{GRanges} object using \code{bsseq}
}
\details{
smooth2txt
}
