% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{smooth2txt}
\alias{smooth2txt}
\title{smooth2txt}
\usage{
smooth2txt(df = df, txt = txt)
}
\arguments{
\item{df}{Data frame}

\item{txt}{Character string of save file name}
}
\value{
Saves a text file
}
\description{
Save smoothed methylation values as a text file
}
