% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothHeatmap.R
\name{smoothHeatmap}
\alias{smoothHeatmap}
\title{smoothHeatmap}
\usage{
smoothHeatmap(regions = sigRegions, bsseq = bs.filtered.bsseq,
  groups = bs.filtered.bsseq \%>\% pData() \%>\% dplyr::as_tibble() \%>\%
  dplyr::pull(!!testCovariate), ...)
}
\arguments{
\item{regions}{\code{GRanges} object of regions to plot a heatmap for}

\item{bsseq}{Smoothed \code{bsseq} object}

\item{groups}{Ordered test covariate information for each sample}

\item{...}{Additional arguments passed onto \code{heatmap.3()}}
}
\value{
An image of the heatmap
}
\description{
Plot a heatmap of normalized individual smoothed methylation value z scores for selected regions (i.e. significant DMRs)
The plotted values are not typical Z-scores, but rather a Z-score of %mCG/CG - mean,
 which is data visualization trick that allows you to focus on the variance
}
\references{
\url{https://sebastianraschka.com/Articles/heatmaps_in_r.html}

\url{https://raw.githubusercontent.com/obigriffith/biostar-tutorials/master/Heatmaps/heatmap.3.R}

\url{https://www.biostars.org/p/18211/}
}
