% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothHeatmap.R
\name{smoothHeatmap}
\alias{smoothHeatmap}
\title{smoothHeatmap}
\usage{
smoothHeatmap(regions = sigRegions, bsseq = bs.filtered.bsseq,
  groups = as.data.frame(pData(bs.filtered.bsseq)) \%>\%
  pull(!!testCovariate),
  out = "sig_individual_smoothed_DMR_methylation.txt", ...)
}
\arguments{
\item{regions}{Genomic ranges object}

\item{bsseq}{Smoothed bsseq object}

\item{groups}{Ordered test covariate information for each sample}

\item{out}{Name of the text file to save in quotations}

\item{names}{Ordered sample names}
}
\value{
Saves a pdf image of the heatmap
}
\description{
Plot a heatmap of individual smoothed methylation value z scores for selected regions (i.e. significant DMRs)
}
