% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tidyDMRs}
\alias{tidyDMRs}
\title{tidyDMRs}
\usage{
tidyDMRs(regions = sigRegionsAnno)
}
\arguments{
\item{regions}{A \code{ChIPseeker csAnno} peak object of DMRs or background regions from \code{dmrseq::dmrseq()}}
}
\value{
A \code{tibble} of annotated regions
}
\description{
Tidy DMRs or background regions from \code{dmrseq::dmrseq()} that have been annotated using \code{ChIPseeker}
}
